+++ src/components/hof/AnonymousHero.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 export default ({ onLogin }) => (
   <div className="hero-card">
     <div className="header card-block1">
-      <img src="svg/square-logo.svg" width="100" height="100" alt="logo" />
+      <img src="/svg/square-logo.svg" width="100" height="100" alt="logo" />
       <div className="header-text">
         <p>Who will be the next one...</p>
         <div className="name">Maybe you?</div>